+++ src/main/java/ru/skypro/lessons/springboot/springf/service/EmployeeService.java
@@ -40,21 +40,6 @@ public interface EmployeeService {
      */
     List<EmployeeFullInfo> getFull();
 
-    /**
-     * POST  принимать на вход файл JSON
-     */
-    void postJsonFileEmployeeRead(MultipartFile file) throws IOException;
-    /**
-     * POST формировать  статистикой по отделам:
-     */
-//    int generateReport() ;
 
-    /**
-     *GET
-     *  находить и возвращать созданный ранее файл в формате JSON по идентификатору.
-     */
-    Optional<Report>  generateReportId(Integer id);
-/////////////////////////////var2//////////////////////////////////////////
-    public int saveJsonPath() ;
 
 }
