+++ src/main/java/ru/skypro/lessons/springboot/springf/service/EmployeeServiceImpl.java
@@ -41,9 +41,7 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     private final EmployeeRepository employeeRepository;
     private final PagingEmployeeRepository pagingEmployeeRepository;
-    private final EmployeeMaper employeeMaper;
-    private final ObjectMapper objectMapper;
-    private final ReportRepository reportRepository;
+
 
 
     /**
@@ -92,98 +90,10 @@ public class EmployeeServiceImpl implements EmployeeService {
         return employeePage.stream().toList();
     }
 
-    /**
-     * POST  принимать на вход файл JSON,
-     * Все сотрудники из загружаемого файла должны быть сохранены в базе данных.
-     */
 
 
-    @Override
-    public void postJsonFileEmployeeRead(MultipartFile file) {
-        writeToFile(file);
-        String filePath = "test.json";
-        ObjectMapper objectMapper = new ObjectMapper();
-
-        try {
-
-            objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-            List<EmployeeDTO> listEmployeeDto = objectMapper.readValue(Paths.get(filePath).toFile(), new TypeReference<>() {
-            });
-            System.out.println(listEmployeeDto);
-            listEmployeeDto.stream()
-                    .map(employeeMaper::toEntity)
-                    .forEach(employeeRepository::save);
-        } catch (IOException ex) {
-            ex.printStackTrace();
-        }
-    }
 
 
-    /**
-     * POST формировать  статистикой по отделам:
-     */
-//    public int generateReport() {
-//        List<ReportDTO> report = reportRepository.buildReport();
-//        try {
-//            String statistic = objectMapper.writeValueAsString(report);
-//
-//            Report reportEntity = new Report();
-//            reportEntity.setData(statistic);
-//            return reportRepository.save(reportEntity).getReportId();
-//        } catch (JsonProcessingException e) {
-//            throw new IllegalStateException("Error", e);
-//        }
-//
-//    }
 
-
-    /**
-     * GET
-     * находить и возвращать созданный ранее файл в формате JSON по идентификатору.
-     */
-//    @Override
-//    public Optional<Report> generateReportId(Integer id) {
-//        Optional<Report> reportsfindById = reportRepository.findById(id);
-//        return reportsfindById;
-//    }
-
-/////////////////////////////////////////////var2///////////////////////////////////
-
-
-    /**
-     * Сформированный отчет необходимо далее сохранить в виде файла на вашем компьютере.
-     * В базу данных вы должны сохранять путь к сформированному отчету.
-     * Метод возвращает целочисленный идентификатор строки в таблице report
-     */
-    @Override
-    public int saveJsonPath() {
-        List<ReportDTO> report = reportRepository.buildReport();
-        String filePath = "testStatistic.json";
-
-        try {
-            String statistic = objectMapper.writeValueAsString(report);
-            System.out.println(statistic);
-            saveStatisticJsonFile(statistic,filePath);
-
-            Report reportPath = new Report();
-
-            reportPath.setFilePath(String.valueOf(Paths.get(filePath)));
-
-            return reportRepository.save(reportPath).getReportId();
-        } catch (JsonProcessingException e) {
-            throw new IllegalStateException("Error", e);
-        }
-    }
-    /**
-     * GET-запрос
-     * localhost:8080/report/{id}
-     *  должен находить и возвращать созданный ранее файл в формате json по переданному уникальному идентификатору,
-     *  используя сохраненный ранее в базу данных путь в файловой системе.
-     */
-    @Override
-    public Optional<Report> generateReportId(Integer id) {
-        Optional<Report> reportsfindById = reportRepository.findById(id);
-        return reportsfindById;
-    }
 }
